+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
 ## master (unreleased)
+* Adds Letter Opener generator. ([@coolprobn][])
+* Adds Whenever generator. ([@aadil][])
+* Adds Rswag generator. ([@coolprobn][])
+
+## 0.12.0 (May 8th, 2023)
 * Adds Flipper generator. ([@abhaynikam][])
 * Adds FactoryBot generator. ([@abhaynikam][])
 * Adds Faker generator. ([@abhaynikam][])
@@ -9,7 +14,6 @@
 * Updates Bootstrap generator for supporting Bootstrap 5 with popper.js. ([@abhaynikam][])
 * Adds Overcommit generator for RuboCop. ([@coolprobn][])
 * Updates RuboCop generator to add rubocop-rake extension and support test frameworks ([@coolprobn][])
-* Adds Rswag generator. ([@coolprobn][])
 
 ## 0.11.0 (June 23rd, 2021)
 * Adds Stimulus generator. ([@abhaynikam][])
@@ -76,3 +80,4 @@
 [@jamesglover]: https://github.com/JamesGlover
 [@luathn]: https://github.com/luathn
 [@coolprobn]: https://github.com/coolprobn
+[@aadil]: https://github.com/AdilRT